club-admin/voting-polls: added voting polls for track club elections
Added voting polls that are using 7 digit voting codes for the voting.
This functionality ensures vote secret by recording only the time code
was used.

The voting logic is placed in the voting_poll cloud functions.

Currently voting polls are configured manually in the system and in
future versions this could be enhanced by using admin UI.